Implemented processor to fetch Coldfront allocation data
Authentication to Coldfront must be done through Keycloak
by providing the client id and secret. These can be set
through env vars (`KEYCLOAK_CLIENT_ID` and `KEYCLOAK_CLIENT_SECRET`).
A Keycloak and Coldfront url can optionally be provided.
They default to the ones hosted by the MGHPCC.

The processor is responsible for adding the PI email
and institute code to invoices.
If a Coldfront allocation for an invoice project could
not be found, and this project also happens to be
billable, a warning will be logged.

Test cases have been added.